Rework the Page Build Action
Instead relying on the Docker image being built in the pages-gem repo,
build our own image here that uses the github-pages gem as a dependency.

Also replace Dockerfile reference with URL in the action metadata file.
In order to support images with different tags, replace the existing
action.yml with a templated file by which the system can generate an
action.yml at build time.